Split Choice component into two wrappers for better flow typing
